+++ packages/vital-card/package.json
@@ -34,6 +34,7 @@
     "@bodiless/components-ui": "^1.0.0-beta.18",
     "@bodiless/fclasses": "^1.0.0-beta.18",
     "@bodiless/hydration": "^1.0.0-beta.18",
+    "@bodiless/vital-buttons": "^1.0.0-beta.18",
     "@bodiless/vital-editors": "^1.0.0-beta.18",
     "@bodiless/vital-elements": "^1.0.0-beta.18",
     "@bodiless/vital-image": "^1.0.0-beta.18",
