+++ sites/vital-demo/src/data/pages/styleguide/card/hero-card$buttontext.json
@@ -1,0 +1,3 @@
+{
+  "text": "asf"
+}
