+++ sites/vital-demo/src/data/pages/styleguide/card/hero-card-primary-button$buttontext.json
@@ -1,0 +1,3 @@
+{
+  "text": "Learn More"
+}
