+++ sites/vital-demo/src/data/pages/styleguide/card/hero-card-primary-button$title.json
@@ -1,0 +1,3 @@
+{
+  "text": "Lorem ipsum"
+}
