+++ sites/vital-demo/src/data/pages/styleguide/card/hero-card.json
@@ -1,3 +1,3 @@
 {
-  "href": "https://jnj.com"
+  "href": "/"
 }
